week2/project/eventms_psycopg2/requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
Flask-Login==0.6.3
python-dotenv==1.0.1
